+++ src/main/java/com/dotweblabs/gwt/client/Parse.java
@@ -39,6 +39,7 @@ import java.util.Iterator;
 public class Parse {
 
     public static class Users {
+        public static String sessionToken = null;
         public static void signup(ParseObject user, final AsyncCallback<ParseResponse> callback) {
             Shape.post(Parse.SERVER_URL + "users")
                     .header("X-Parse-Application-Id", X_Parse_Application_Id)
@@ -56,6 +57,7 @@ public class Parse {
                             ParseResponse response = ParseResponse.parse(s);
                             if(response.get("sessionToken") != null) {
                                 //createdAt, objectId, sessionToken
+                                sessionToken = response.get("sessionToken").isString().stringValue();
                                 callback.onSuccess(response);
                             } else {
                                 HttpRequestException ex
@@ -64,13 +66,14 @@ public class Parse {
                             }
                         }
                     });
-        }
+    }
         public static void login(String username, String password, final AsyncCallback<ParseResponse> callback) {
             String param = "username=" + username + "&" + "password=" + password;
             Shape.get(Parse.SERVER_URL + "login?" + URL.encode(param))
                     .header("X-Parse-Application-Id", X_Parse_Application_Id)
                     .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)
                     .header("X-Parse-Master-Key", X_Parse_Master_Key)
+                    .header("X-Parse-Revocable-Session", "1")
                     .asJson(new AsyncCallback<String>() {
                         @Override
                         public void onFailure(Throwable throwable) {
@@ -82,12 +85,42 @@ public class Parse {
                             ParseResponse response = ParseResponse.parse(s);
                             if(response.get("sessionToken") != null) {
                                 //createdAt, objectId, sessionToken
+                                sessionToken = response.get("sessionToken").isString().stringValue();
                                 callback.onSuccess(response);
                             } else {
                                 HttpRequestException ex
                                         = new HttpRequestException(response.getErrorMessage(), response.getErrorCode());
                                 callback.onFailure(ex);
                             }
+                        }
+                    });
+        }
+        public static void logout(final AsyncCallback<ParseResponse> callback) {
+            Shape.post(Parse.SERVER_URL + "logout")
+                    .header("X-Parse-Application-Id", X_Parse_Application_Id)
+                    .header("X-Parse-REST-API-Key", X_Parse_REST_API_Key)
+                    .header("X-Parse-Master-Key", X_Parse_Master_Key)
+                    .header("X-Parse-Session-Token", sessionToken)
+                    .asJson(new AsyncCallback<String>() {
+                        @Override
+                        public void onFailure(Throwable throwable) {
+                            HttpRequestException ex = (HttpRequestException) throwable;
+                            callback.onFailure(ex);
+                        }
+                        @Override
+                        public void onSuccess(String s) {
+                            ParseResponse response = ParseResponse.parse(s);
+                            sessionToken = null;
+                            callback.onSuccess(response);
+                            /*
+                            if(response.get("sessionToken") != null) {
+                                //createdAt, objectId, sessionToken
+                                callback.onSuccess(response);
+                            } else {
+                                HttpRequestException ex
+                                        = new HttpRequestException(response.getErrorMessage(), response.getErrorCode());
+                                callback.onFailure(ex);
+                            }*/
                         }
                     });
         }